refactor: service의 openApi 담당 계층 분리
